requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiosqlite>=0.19.0
google-genai>=0.1.0
